Pulled Provenance out into its own file.
